Move blueprint stuff to base contract
