+++ boa/contracts/base_evm_contract.py
@@ -132,3 +132,20 @@ class BoaError(Exception):
         call_tree = str(self.call_trace)
         ledge = "=" * 72
         return f"\n{ledge}\n{call_tree}\n{ledge}\n\n{ret}"
+
+
+DEFAULT_BLUEPRINT_PREAMBLE = b"\xFE\x71\x00"
+
+
+def generate_blueprint_bytecode(
+    contract_bytecode: bytes, blueprint_preamble: bytes = DEFAULT_BLUEPRINT_PREAMBLE
+):
+    if blueprint_preamble is None:
+        blueprint_preamble = b""
+    blueprint_bytecode = blueprint_preamble + contract_bytecode
+
+    # the length of the deployed code in bytes
+    len_bytes = len(blueprint_bytecode).to_bytes(2, "big")
+    deploy_bytecode = b"\x61" + len_bytes + b"\x3d\x81\x60\x0a\x3d\x39\xf3"
+
+    return deploy_bytecode + blueprint_bytecode
